DATACASS-146 - Polishing.
Reorder methods to match interface ordering. Remove superfluous null check.

Refactor common code shared by CassandraQuery classes into CassandraRepositoryQuerySupport.
