+++ spring-data-cassandra/src/main/java/org/springframework/data/cassandra/repository/query/CassandraParametersParameterAccessor.java
@@ -54,15 +54,6 @@ public class CassandraParametersParameterAccessor extends ParametersParameterAcc
 
 	/*
 	 * (non-Javadoc)
-	 * @see org.springframework.data.cassandra.repository.query.CassandraParameterAccessor#findCassandraType(int)
-	 */
-	@Nullable
-	public CassandraType findCassandraType(int index) {
-		return getParameters().getParameter(index).getCassandraType();
-	}
-
-	/*
-	 * (non-Javadoc)
 	 * @see org.springframework.data.cassandra.repository.query.CassandraParameterAccessor#getDataType(int)
 	 */
 	@Override
@@ -76,11 +67,11 @@ public class CassandraParametersParameterAccessor extends ParametersParameterAcc
 
 	/*
 	 * (non-Javadoc)
-	 * @see org.springframework.data.repository.query.ParametersParameterAccessor#getParameters()
+	 * @see org.springframework.data.cassandra.repository.query.CassandraParameterAccessor#findCassandraType(int)
 	 */
-	@Override
-	public CassandraParameters getParameters() {
-		return (CassandraParameters) super.getParameters();
+	@Nullable
+	public CassandraType findCassandraType(int index) {
+		return getParameters().getParameter(index).getCassandraType();
 	}
 
 	/*
@@ -90,6 +81,15 @@ public class CassandraParametersParameterAccessor extends ParametersParameterAcc
 	@Override
 	public Class<?> getParameterType(int index) {
 		return getParameters().getParameter(index).getType();
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * @see org.springframework.data.repository.query.ParametersParameterAccessor#getParameters()
+	 */
+	@Override
+	public CassandraParameters getParameters() {
+		return (CassandraParameters) super.getParameters();
 	}
 
 	/*
